# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Travelode API
host: 127.0.0.1:28825
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /travelode:
    # binds a127 app logic to a route
    x-swagger-router-controller: travelode

    get:
      tags:
        - Travelode
      summary: Returns travelode list
      operationId: findTravelodes
      parameters:
        - name: page
          in: query
          description: The page number to get. Default to 0.
          required: false
          type: string
        - name: size
          in: query
          description: The size of the page to get. Default to 50.
          required: false
          type: string
        - name: title
          in: query
          description: Find by travelode title
          required: false
          type: string
        - name: userId
          in: query
          description: Find by userId
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TravelodeListResponse"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      tags:
        - Travelode
      summary: Create new travelode
      operationId: createTravelode
      consumes:
        - multipart/form-data
      produces:
          - application/json
      parameters:
        - in: formData
          name: title
          description: Title of the Travelode
          required: true
          type: string
        - in: formData
          name: description
          description: Description of the Travelode
          required: false
          type: string
        - in: formData
          name: userId
          description: Creator / owner of the Travelode
          required: true
          type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/CreatedResponse"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /travelode/{id}:
    x-swagger-router-controller: travelode

    get:
      tags:
        - Travelode
      summary: Returns a single travelode details
      operationId: getTravelodeById
      parameters:
        - name: id
          in: path
          description: The Travelode Id
          required: true
          type: integer
          minimum: 1
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TravelodeListResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Travelode
      summary: Update a travelode
      operationId: updateTravelodeById
      parameters:
        - name: id
          in: path
          description: The travelode id to update
          required: true
          type: integer
          minimum: 1
        - name: travelodeData
          in: body
          description: The travelode data to update
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Travelode
      summary: Delete a travelode
      operationId: deleteTravelodeById
      parameters:
        - name: id
          in: path
          description: The travelode id to delete
          required: true
          type: integer
          minimum: 1
      responses:
        "204":
          description: Travelode is removed
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  CreatedResponse:
    required:
      - message
      - id
    properties:
      message:
        type: string
      id:
        type: number
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Travelode:
    type: object
    properties:
      id:
        type: number
        description: Travelode Id
      title:
        type: string
        description: Travelode Title
      description:
        type: string
        description: Travelode Description
      userId:
        type: string
        description: Travelode Owner UserId
      created:
        description: Travelode Creation Date Time
        type: string
        format: date-time
      updated:
        description: Travelode Update Date time
        type: string
        format: date-time
  TravelodeListResponse:
    required:
      - travelodes
    properties:
      # The array of travelodes
      travelodes:
        type: array
        items:
          $ref: "#/definitions/Travelode"
...